refactor: 상품 관리 페이지 변수명 및 파일명 수정
